move checkpoints into x-pack core and introduce base classes for data frame tests
